handle date at end of month
move goal into model module
