+++ src/templates/accounts/form.turbo.html
@@ -1,7 +1,5 @@
 <turbo-stream action="update" target="errors">
   <template>
-    {% if errors %}
-      <p>{{errors}}</p>
-    {% endif %}
+    {% if errors %}{{errors}}{% endif %}
   </template>
 </turbo-stream>
