+++ src/templates/dashboard.html
@@ -5,7 +5,6 @@
 {% set deg_value = 1.0 - ((78.0 - remaining) / 20.0)  %}
 {% set deg = 90 * deg_value %}
 
-
 <div class="flex justify-center">
   <div class="w-64">
   <div class="bg-blue-400 aspect-[2] rounded-t-full relative flex items-center justify-center overflow-hidden">
